% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportHTML.R
\name{can.interact}
\alias{can.interact}
\title{Identify if a plot can be interactive}
\usage{
can.interact(x)
}
\arguments{
\item{x}{the plot object (either from iNZightPlots or iNZightMaps)}
}
\value{
Logical to identify if there is an interactive version
}
\description{
Identify if a plot can be interactive
}
