% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpalettes.R
\name{const_palette_names}
\alias{const_palette_names}
\title{An incorrectly spelled function - deprecated}
\usage{
const_palette_names()
}
\value{
a list of continuous colour palettes
}
\description{
This function was misspelled in earlier versions
and has been corrected to \code{cont_palette_names},
which should be used instead.
}
\seealso{
\code{\link{cont_palette_names}}
}
