% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{construct_call}
\alias{construct_call}
\title{Construct plot call from settings list}
\usage{
construct_call(
  settings,
  vartypes,
  data = quote(.dataset),
  design = quote(.design),
  what = c("plot", "summary", "inference")
)
}
\arguments{
\item{settings}{a list of plot settings, similar to \code{inzpar()}}

\item{vartypes}{a list of variables types (numeric, factor)}

\item{data}{a data set to pass to the call}

\item{design}{a survey design (can be NULL)}

\item{what}{the type of call to produce}
}
\value{
a plot/summary/inference call
}
\description{
Construct plot call from settings list
}
