% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert.to.factor}
\alias{convert.to.factor}
\title{Convert to Factor}
\usage{
convert.to.factor(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a factor vector
}
\description{
Convert a numeric variable in to a factor with four levels.
}
\author{
tell029
}

