% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{create}
\alias{create}
\title{Method for creating plots for iNZight}
\usage{
create(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
a plotable object
}
\description{
Create a Plot Object
}
\author{
Tom Elliott
}
