% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exploreAllPlots}
\alias{exploreAllPlots}
\title{Explore all Univariate Plots}
\usage{
exploreAllPlots(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Allows easy viewing of every variable in the data set.
}
\author{
tell029
}

