% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exploreAllSummaries}
\alias{exploreAllSummaries}
\title{Explore all Univariate Summaries}
\usage{
exploreAllSummaries(data)
}
\arguments{
\item{data}{a data set}
}
\value{
allSummaries object
}
\description{
Allows easy access to a summary for every variable in the data set.
}
\author{
tell029
}

