% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSVG.R
\name{exportSVG}
\alias{exportSVG}
\title{Export iNZightPlots as an SVG}
\usage{
exportSVG(x, file)
}
\arguments{
\item{x}{iNZight plot object or function that captures iNZight environment}

\item{file}{Name of temporary svg file generated (by default: 'inzightplot.svg')}
}
\value{
Opens up an SVG file of \code{x} with filename \code{file} in a web browser
}
\description{
\code{exportSVG} is designed to export the iNZight plot as a temporary SVG that is opened in a web browser.
The iNZightPlot must be drawn to a graphics device before exporting can occur.
}
\author{
Yu Han Soh
}
