% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format.out_node}
\alias{format.out_node}
\title{format method for nodes}
\usage{
\method{format}{out_node}(x, format = c("plain", "html"), width = 100L, ...)
}
\arguments{
\item{x}{a node object to be formatted}

\item{format}{the desired output format ("plain" or "html" at present)}

\item{width}{the width of the output}

\item{...}{additional arguments}
}
\description{
format method for nodes
}
