% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzinference.R
\name{inzinference}
\alias{inzinference}
\alias{inzinference.formula}
\title{iNZight Inference Method}
\usage{
inzinference(x, ..., env = parent.frame())

\method{inzinference}{formula}(
  x,
  data = NULL,
  design = NULL,
  type = c("conf", "comp"),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{x}{An object}

\item{...}{additional arguments for methods}

\item{env}{an environment to evaluate things}

\item{data}{Dataset to plotq}

\item{design}{A survey design to use}

\item{type}{Type type of inference to obtain, one of 'conf' or 'comp'
for confidence intervals and comparison intervals, respectively
(currently ignored).}
}
\value{
The output depends on the type of input, and consists of a inference object with a \code{print} method.
}
\description{
A generic function used to generate inferential information for objects within the iNZight ecosystem.
}
\section{Methods (by class)}{
\itemize{
\item \code{inzinference(formula)}: Wrapper for getPlotSummary to obtain inference information about a plot

}}
