% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzsummary.R
\name{inzsummary}
\alias{inzsummary}
\alias{inzsummary.formula}
\title{iNZight Summary Method}
\usage{
inzsummary(x, ..., env = parent.frame())

\method{inzsummary}{formula}(x, data = NULL, design = NULL, ..., env = parent.frame())
}
\arguments{
\item{x}{An object}

\item{...}{an environment to evaluate things}
}
\value{
The output depends on the type of input, and consists of a summary object with a \code{print} method.
}
\description{
A generic function used to summarize objects within the iNZight ecosystem.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Wrapper for getPlotSummary to obtain summary information about a plot
}}

