% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{mend_call}
\alias{mend_call}
\title{Mend a plot call based on valid parameters}
\usage{
mend_call(call, data, design_name, plot)
}
\arguments{
\item{call}{a plot call string, or expression}

\item{data}{the dataset}

\item{design_name}{name of the design, if any}

\item{plot}{the result of \code{inzplot}, \code{inzsummary}, or \code{inzinference}}
}
\value{
a plot call with extraneous arguments removed
}
\description{
Mend a plot call based on valid parameters
}
