% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inzplotoutput.R
\name{print.inzplotoutput}
\alias{print.inzplotoutput}
\alias{plot.inzplotoutput}
\title{Print an iNZight plot object}
\usage{
\method{print}{inzplotoutput}(x, ...)

\method{plot}{inzplotoutput}(x, y, ...)
}
\arguments{
\item{x}{an \code{inzplotoutput} object}

\item{...}{additional arguments (ignored)}

\item{y}{ignored}
}
\value{
\code{x}, invisibly
}
\description{
Draws an \code{inzplotoutput} created by \code{\link{iNZightPlot}}.
Called automatically when the object is printed; call explicitly to redraw.
}
