% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_controls.R
\name{snz_privacy_controls}
\alias{snz_privacy_controls}
\title{Statistics New Zealand Privacy Controls}
\usage{
snz_privacy_controls(type = c("survey"), weighted = type == "survey", ...)
}
\arguments{
\item{type}{the type of data, used to specify the correct rules. Currently only survey (4.0.1) data is supported.}

\item{weighted}{logical indicating if the results are a weighted survey design or not.}

\item{...}{additional arguments, used to override defaults}
}
\value{
a list of privacy control rules
}
\description{
Based off Microdata Output Guide 2020 v5-1
}
