% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{convert.to.factor}
\alias{convert.to.factor}
\title{Convert to Factor}
\usage{
convert.to.factor(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a factor variable
}
\description{
Convert a numeric variable in to a factor with four levels.
}
\examples{
f <- convert.to.factor(runif(100, 0, 10))
levels(f)
}
\author{
Tom Elliott
}
