% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{create}
\alias{create}
\title{Create plots for iNZight}
\usage{
create(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
an iNZight plot object with class determined by data type
}
\description{
Create a Plot Object
}
\details{
This create method is to be used by packages extending 'iNZightPlots',
and should not be used by users. The resulting object should have
an associated \code{plot} method.
}
\author{
Tom Elliott
}
