% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreAllPlots.R
\name{exploreAllPlots}
\alias{exploreAllPlots}
\title{Explore all Univariate Plots}
\usage{
exploreAllPlots(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Allows easy viewing of every variable in the data set.
The user will be prompted to see the next variable.
}
\examples{
if (interactive())
    exploreAllPlots(iris)

}
\author{
Tom Elliott
}
