% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreAllSummaries.R
\name{exploreAllSummaries}
\alias{exploreAllSummaries}
\alias{print.allSummaries}
\title{Explore all Univariate Summaries}
\usage{
exploreAllSummaries(data, ...)

\method{print}{allSummaries}(x, ...)
}
\arguments{
\item{data}{a data set}

\item{...}{additional arguments passed to getPlotSummary()}

\item{x}{an \code{allSummaries} object}
}
\value{
allSummaries object
}
\description{
Allows easy access to a summary for every variable in the data set.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: print method for allSummaries object
}}

\author{
Tom Elliott
}
