% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportHTML.R
\name{exportHTML}
\alias{exportHTML}
\alias{exportHTML.function}
\alias{exportHTML.inzplotoutput}
\title{ExportHTML}
\usage{
exportHTML(x, file, data, extra.vars, ...)

\method{exportHTML}{function}(x, file = "index.html", data = NULL,
  extra.vars = NULL, width = dev.size()[1], height = dev.size()[2], ...)

\method{exportHTML}{inzplotoutput}(x, file = "index.html", data = NULL,
  extra.vars = NULL, ...)
}
\arguments{
\item{x}{An iNZight plot object or function (such as updatePlot) that captures iNZight environment}

\item{file}{Name of temporary HTML file generated
Additional parameters for scatterplots only:}

\item{data}{dataset/dataframe that you wish to investigate and export more variables from}

\item{extra.vars}{extra variables specified by the user to be exported}

\item{...}{extra arguments}

\item{width}{the desired width of the SVG plot}

\item{height}{the desired height of the SVG plot}
}
\value{
Opens up an HTML file of \code{x} with filename \code{file} in the browser (best performance on Chrome/Firefox)
}
\description{
\code{exportHTML} is designed to export the iNZight plot as a dynamic, interactive HTML page.
Currently only handles single panel plots. Coloured hex plots are currently not available yet.
}
\section{Methods (by class)}{
\itemize{
\item \code{function}: method for an iNZightPlot-generating function

\item \code{inzplotoutput}: method for output from iNZightPlot
}}

\examples{
\dontrun{
x <- iNZightPlot(Petal.Width, Petal.Length, data = iris, colby = Species)
exportHTML(x, "index.html")

#to export more variables for scatterplots:
 exportHTML(x, "index.html", data = iris, extra.vars = c("Sepal.Length", "Sepal.Width"))
}

}
\author{
Yu Han Soh
}
