% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportHTML.R
\name{exportHTML}
\alias{exportHTML}
\alias{exportHTML.function}
\alias{exportHTML.ggplot}
\alias{exportHTML.inzplotoutput}
\title{ExportHTML}
\usage{
exportHTML(x, file, data, local = FALSE, extra.vars, ...)

\method{exportHTML}{`function`}(
  x,
  file = "index.html",
  data = NULL,
  local = FALSE,
  extra.vars = NULL,
  width = dev.size()[1],
  height = dev.size()[2],
  ...
)

\method{exportHTML}{ggplot}(
  x,
  file = "index.html",
  data = NULL,
  local = FALSE,
  extra.vars = NULL,
  mapObj,
  ...
)

\method{exportHTML}{inzplotoutput}(
  x,
  file = "index.html",
  data = NULL,
  local = FALSE,
  extra.vars = NULL,
  ...
)
}
\arguments{
\item{x}{An iNZight plot object that captures iNZight environment}

\item{file}{Name of temporary HTML file generated}

\item{data}{dataset/dataframe that you wish to investigate and export more variables from}

\item{local}{Logical for creating local files for offline use (default to false)
Additional parameters for scatterplots and dotplots only:}

\item{extra.vars}{extra variables specified by the user to be exported}

\item{...}{extra arguments}

\item{width}{the desired width of the SVG plot}

\item{height}{the desired height of the SVG plot}

\item{mapObj}{iNZightMap object (from iNZightMaps)}
}
\value{
HTML file of \code{x} with filename \code{file} in the browser
}
\description{
\code{exportHTML} is designed to export the iNZight plot as a dynamic, interactive HTML page.
Currently only handles single panel plots. Coloured hex plots are currently not available yet.
}
\section{Methods (by class)}{
\itemize{
\item \code{function}: method for an iNZightPlot-generating function

\item \code{ggplot}: method for iNZightMaps or other supported ggplot graphs

\item \code{inzplotoutput}: method for output from iNZightPlot
}}

\examples{
\dontrun{
x <- iNZightPlot(Petal.Width, Petal.Length, data = iris, colby = Species)
exportHTML(x, "index.html")

#to export more variables for scatterplots:
 exportHTML(x, "index.html", data = iris, extra.vars = c("Sepal.Length", "Sepal.Width"))
}

}
\author{
Yu Han Soh
}
