% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iNZightPlot}
\alias{iNZightPlot}
\title{iNZight Plot}
\usage{
iNZightPlot(x, y = NULL, g1 = NULL, g1.level = NULL, g2 = NULL,
  g2.level = NULL, varnames = list(), colby = NULL, sizeby = NULL,
  data = NULL, design = NULL, freq = NULL, missing.info = TRUE,
  xlab = varnames$x, ylab = varnames$y, new = TRUE, inzpars = inzpar(),
  layout.only = FALSE, plot = TRUE, df, env = parent.frame(), ...)
}
\arguments{
\item{x}{a vector (numeric or factor), or the name of a column in the supplied
\code{data} or \code{design} object}

\item{y}{a vector (numeric or factor), or the name of a column in the supplied
\code{data} or \code{design} object}

\item{g1}{a vector (numeric or factor), or the name of a column in the supplied
\code{data} or \code{design} object. This variable acts as a subsetting variable.}

\item{g1.level}{the name (or numeric position) of the level of \code{g1} that will be
used instead of the entire data set}

\item{g2}{a vector (numeric or factor), or the name of a column in the supplied
\code{data} or \code{design} object. This variable acts as a subsetting variable, similar to
\code{g1}}

\item{g2.level}{same as \code{g1.level}, however takes the additional value \code{"_MULTI"},
which produces a matrix of \code{g1} by \code{g2}}

\item{varnames}{a list of variable names, with the list named using the appropriate arguments
(i.e., \code{list(x = "height", g1 = "gender")})}

\item{colby}{the name of a variable (numeric or factor) to colour points by. In the
case of a numeric variable, a continuous colour scale is used, otherwise each level of
the factor is assigned a colour}

\item{sizeby}{the name of a (numeric) variable, which controls the size of points}

\item{data}{the name of a data set}

\item{design}{the name of a survey object, obtained from the \code{survey} package}

\item{freq}{the name of a frequency variable if the data are frequencies}

\item{missing.info}{logical, if \code{TRUE}, information regarding missingness is
displayed in the plot}

\item{xlab}{the text for the x-label}

\item{ylab}{the text for the y-label}

\item{new}{logical, used for compatibility}

\item{inzpars}{allows specification of iNZight plotting parameters over multiple plots}

\item{layout.only}{logical, if \code{TRUE}, only the layout is drawn (useful if a
custom plot is to be drawn)}

\item{plot}{logical, if \code{FALSE}, the plot is not drawn (used by \code{summary})}

\item{df}{compatibility argument}

\item{env}{compatibility argument}

\item{...}{additional arguments, see \code{inzpar}}
}
\value{
An \code{inzightplotoutput} object, which contains the information displayed
in the plot
}
\description{
A general plotting function that automatically detects variable type and draws the
appropriate plot.
It also provides facilities to add inference information to plots, colour- and
size-by variables, and can handle survey data.
}
\details{
Some details here ...
}
\author{
tell029
}

