% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpalettes.R
\name{inzpalette}
\alias{inzpalette}
\title{iNZight colour palette}
\usage{
inzpalette(palette, type)
}
\arguments{
\item{palette}{the name of a palette}
}
\value{
a colour palette function, with argument `n`
}
\description{
Used to obtain a colour palette of a given name.
A list of available palettes can be obtained by
`cat_palette_names()` and `cont_palette_names()`.
}
\author{
Tom Elliott
}
