% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpalettes.R
\name{inzpalette}
\alias{inzpalette}
\alias{cat_palette_names}
\alias{cont_palette_names}
\title{iNZight colour palette}
\usage{
inzpalette(palette)

cat_palette_names()

cont_palette_names()
}
\arguments{
\item{palette}{the name of a palette}
}
\value{
a colour palette function with single argument `n`
}
\description{
Used to obtain a colour palette of a given name.
A list of available palettes can be obtained by
`cat_palette_names()` and `cont_palette_names()`.
}
\section{Functions}{
\itemize{
\item \code{cat_palette_names()}: List of categorical colour palettes

\item \code{cont_palette_names()}: List of continuous  colour palettes

}}
\examples{
plot(1:5, pch = 19, col = inzpalette("bright")(5))

# for a list of palette names
cat_palette_names()
cont_palette_names()
}
\author{
Tom Elliott
}
