% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzpar.R
\name{inzpar}
\alias{inzpar}
\title{iNZight Plotting Parameters}
\usage{
inzpar(..., .viridis = requireNamespace("viridis", quietly = TRUE))
}
\arguments{
\item{...}{If arguments are supplied, then these values are set. If left empty, then}

\item{.viridis}{checks if the viridis package is installed; or can be turned off
the default list is returned.}
}
\value{
an object of class \code{inzpar.list}
}
\description{
Plotting parameters for iNZight Plots
}
\details{

}
\examples{
# arguments can be passed directly to \code{iNZightPlot}
iNZightPlot(Sepal.Length, data = iris, col.pt = "red",
    box.col = "blue", box.fill = "green")

# or stored and passed to it (only pars relevant to the current
# plot are used)
mypar <- inzpar(col.pt = "red", box.col = "blue", box.fill = "green",
    trend = "linear", trend.by = TRUE)
inzplot(Sepal.Length ~ Species, data = iris, inzpar = mypar)
iNZightPlot(Sepal.Length, Sepal.Width, data = iris, inzpar = mypar,
    colby = Species)
}
