% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inzpar}
\alias{inzpar}
\title{iNZight Plotting Parameters}
\usage{
inzpar(...)
}
\arguments{
\item{...}{If arguments are supplied, then these values are set. If left empty, then
the default list is returned.}
}
\value{
an object of class \code{inzpar.list}
}
\description{
Plotting parameters for iNZight Plots
}
\details{
A whole suite of parameters that can be used to fine-tune plots obtained from the
\code{iNZightPlot} function.
}
\author{
tell029
}

