% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZPlot.R
\name{iNZPlot}
\alias{iNZPlot}
\title{iNZightPlot Formula Interface}
\usage{
iNZPlot(f, data = NULL, ...)
}
\arguments{
\item{f}{A formula in the form of \code{y ~ x | g}. See Details.}

\item{data}{Dataset to plot}

\item{...}{Any arguments to pass to \code{\link{iNZightPlot}}}
}
\value{
An \code{inzightplotoutput} object, which contains the information
        displayed in the plot
}
\description{
iNZightPlot Formula Interface
}
\details{
\code{iNZPlot} is a simple wrapper around the \code{\link{iNZightPlot}} function. 

There are four options for the formula passed in:

\code{y} will produce a plot of the single variable \code{y}.

\code{y ~ x} will produce a plot of \code{y} against \code{x}.

\code{y ~ x | g1} will produce a plot of \code{y} against \code{x} subset by \code{g1}.

\code{y ~ x | g1 + g2} will produce a plot of \code{y} against \code{x} subset by \code{g1} and \code{g2}.
}
\examples{
data("CO2")
iNZPlot(uptake, data = CO2)
iNZPlot(uptake ~ Treatment, data = CO2)
iNZPlot(uptake ~ Treatment | Type, data = CO2)
iNZPlot(uptake ~ Treatment | Type, 
data = CO2, g1.level = "Quebec")
}
\seealso{
iNZightPlot
}
